+++ src/components/LayerAttribute/HeatmapLayerStyle/index.tsx
@@ -1,15 +1,11 @@
 import classNames from 'classnames';
 import React, { useMemo, memo } from 'react';
-import { Form, FormCollapse, FormItem, Input, NumberPicker, Select, Switch } from '@formily/antd';
+import { Form, FormItem, Input, NumberPicker, Select, Switch } from '@formily/antd';
 import { createSchemaField } from '@formily/react';
 import type { Form as FormInstance } from '@formily/core';
 import { createForm, onFormValuesChange } from '@formily/core';
-import Collapse from '../components/Collapse';
-import FieldSelect from '../components/FieldSelect';
-import ColorPicker from '../components/ColorPicker';
-import RibbonSelect from '../components/RibbonSelect';
-import Slider from '../components/Slider';
-import SliderRange from '../components/SliderRange';
+import { debounce } from 'lodash-es';
+import { FormCollapse, FieldSelect, ColorPicker, RibbonSelect, Slider, SliderRange } from '../components';
 import type { HeatmapLayerStyleAttributeProps } from './types';
 import schema from './schema';
 import { heatmapLayerStyleConfigToFlat, heatmapLayerStyleFlatToConfig } from './helper';
@@ -31,7 +27,6 @@ export const HeatmapLayerStyleAttributeSchemaField: React.FC<Pick<HeatmapLayerSt
           Slider,
           RibbonSelect,
           ColorPicker,
-          Collapse,
           FieldSelect,
           SliderRange,
         },
@@ -51,9 +46,11 @@ export const HeatmapLayerStyleAttribute: React.FC<HeatmapLayerStyleAttributeProp
       const _form = createForm({
         initialValues,
         effects() {
-          onFormValuesChange((formIns: FormInstance<any>) => {
-            props.onChange(heatmapLayerStyleFlatToConfig(formIns.values));
-          });
+          onFormValuesChange(
+            debounce((formIns: FormInstance<any>) => {
+              props.onChange(heatmapLayerStyleFlatToConfig(formIns.values));
+            }, 150),
+          );
         },
       });
