+++ src/components/LayerAttribute/HeatmapLayerStyle/schema.ts
@@ -5,7 +5,7 @@ export default (fieldList: FieldSelectOptionType[] = []) => {
     properties: {
       color: {
         type: 'void',
-        'x-component': 'Collapse',
+        'x-component': 'FormCollapse',
         'x-component-props': {
           ghost: true,
           destroyInactivePanel: true,
@@ -14,7 +14,7 @@ export default (fieldList: FieldSelectOptionType[] = []) => {
         properties: {
           color: {
             type: 'void',
-            'x-component': 'Collapse.CollapsePanel',
+            'x-component': 'FormCollapse.CollapsePanel',
             'x-component-props': {
               header: '热力颜色',
             },
@@ -73,26 +73,27 @@ export default (fieldList: FieldSelectOptionType[] = []) => {
         },
       },
 
-      // 半径
-      radius: {
+      // 热力半径
+      collapseItem_fillSize: {
         type: 'void',
-        'x-component': 'Collapse',
+        'x-component': 'FormCollapse',
         'x-component-props': {
           ghost: true,
           destroyInactivePanel: true,
           defaultActiveKey: [],
         },
         properties: {
-          radius: {
+          fillsize: {
             type: 'void',
-            'x-component': 'Collapse.CollapsePanel',
+            'x-component': 'FormCollapse.CollapsePanel',
             'x-component-props': {
               header: '热力半径',
             },
+
             properties: {
               radius: {
                 type: 'number',
-                title: '半径',
+                title: '热力半径',
                 default: 20,
                 'x-decorator-props': {},
                 'x-decorator': 'FormItem',
@@ -103,29 +104,6 @@ export default (fieldList: FieldSelectOptionType[] = []) => {
                   step: 1,
                 },
               },
-            },
-          },
-        },
-      },
-
-      // 权重
-      collapseItem_fillSize: {
-        type: 'void',
-        'x-component': 'Collapse',
-        'x-component-props': {
-          ghost: true,
-          destroyInactivePanel: true,
-          defaultActiveKey: [],
-        },
-        properties: {
-          fillsize: {
-            type: 'void',
-            'x-component': 'Collapse.CollapsePanel',
-            'x-component-props': {
-              header: '权重',
-            },
-
-            properties: {
               sizeField: {
                 type: 'string',
                 title: '权重字段',
